pyproject.toml
[tool.poetry]
name = "OpenOligo"
version = "0.1.6"
description = "An open-source platform for programmatically interacting with and managing Nucleic acid sequences synthesis processes."
license = "Apache-2.0"
authors = ["Satyam Tiwary <satyam@technoculture.io>"]
readme = "README.md"
keywords = ["DNA", "synthesis", "genetics", "open-source"]

[tool.poetry.scripts]
oligo-server = 'scripts.server:main'
oligo-runner = 'scripts.runner:main'

[tool.poetry.dependencies]
python = "^3.9"
rich = "^13.4.2"
python-dotenv = "^1.0.0"
tqdm = "^4.65.0"
fastapi = "^0.98.0"
pdocs = "^1.2.0"
uvicorn = "^0.22.0"
"RPi.GPIO" = { version = "^0.7.1", optional = true }
Adafruit_BBIO = { version = "^1.2.0", optional = true }
aerich = "^0.7.1"
types-tqdm = "^4.65.0.1"
tortoise-orm = "^0.18.1"
httpx = "^0.24.1"
sh = "^2.0.4"
anyio = "^3.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
flake8 = "^3.9.2"
black = "^23.3.0"
pylint = "^2.17.4"
mypy = "^1.3.0"
isort = "^5.12.0"
coverage-badge = "^1.1.0"
jupyter = "^1.0.0"
jupyterlab = "^4.0.2"
setuptools = "^68.0.0"
wheel = "^0.40.0"
twine = "^4.0.2"
# ansible = "^8.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100

[tool.pytest.ini_options]
addopts = "--cov=openoligo --cov-config=pyproject.toml --cov-report=term-missing --cov-fail-under=70"

[tool.coverage.run]
source = ["openoligo"]
#omit = ["**/types.py"]

[tool.poetry.extras]
rpi = ["RPi.GPIO"]
bb = ["Adafruit_BBIO"]
